+++ src/app/core/models/constants.model.ts
@@ -1,0 +1,5 @@
+export interface ConstantsModel {
+  appName: string;
+  sunColor: string;
+  yearCreated: number;
+}