+++ src/app/core/services/config-options.service.ts
@@ -1,0 +1,13 @@
+import { ConfigModel } from '../models/config.model';
+
+export class ConfigOptionsService {
+  private config: ConfigModel;
+
+  getConfig() {
+    return this.config;
+  }
+
+  setConfig(config: Partial<ConfigModel>) {
+    this.config = { ...this.config, ...config };
+  }
+}