+++ src/app/core/services/constants.service.ts
@@ -1,0 +1,7 @@
+import { ConstantsModel } from '../models/constants.model';
+
+export const ConstantsService: ConstantsModel = {
+  appName: 'Shop Market',
+  sunColor: 'Yellow',
+  yearCreated: 2021
+}