+++ src/app/core/tokens/constants.token.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { ConstantsModel } from '../models/constants.model';
+
+export const constantsToken = new InjectionToken<ConstantsModel>('constants token');